Media server discovery test (3.1 in bc spreadsheet)
It says media server should be found by looking for SRV record.
Yet, buddycloud.org doesn't a SRV pointing to media server.

I notice we can reach it by xep_0300 querying the XMPP server
to find its items. An item representing the mediaserver is one of the
results. So that's what this test is doing as of now.