Many things.
First, if a test fails not because you haven't acomplished what
it asks, but because you haven't acomplished something this test relies on
e.g. another test that this one depends upon, this test is skipped
and the user is informed so.

Second, many enhancements on output of first 2 tests.

Third, couple typos fixes.