Update visibility_utils.py
Fixes the problem seen below:

Exception ValueError with message "No JSON object could be decoded" is being raised when running all the visibility tests.

The root of the problem is that queries for a node's metadata are returning status 200 even if that node doesn't exist.

It means it's deffo a bug somewhere else, as the specs (https://buddycloud.org/wiki/buddycloud_HTTP_API#.2F:channel.2Fmetadata.2F:node) say it should return status 404 in that case.

Now, to fix the issue, the tests impacted by this commit actually rely on checking if the response's content is actually in valid JSON format before considering the query for metadata a success. (Maybe it should also check for the expected keys within that JSON object).