Client cancelling launch doesn't terminate process
For now, if the client cancels a run, it won't cause
the server to terminate the run process so that we can make
sure that the teardown processes will eventually run
(thus deleting all residual test channels created meanwhile testing).

But what needs to be done later is a better isolation
between tests, set up procedures and teardown procedures so
that if the client cancels the run, we terminate the testing
procedures and just have the teardown procedures run normally.